backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Adrià Gomis", email = "gomisroca@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "celery>=5.6.3",
    "fastapi[standard]>=0.136.1",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.14.0",
    "pydantic[email]>=2.13.3",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "resend>=2.30.0",
    "sqlalchemy>=2.0.49",
    "stripe>=15.1.0",
    "uvicorn[standard]>=0.46.0",
]
[tool.fastapi]
entrypoint = "backend.main:app"

[project.scripts]
backend = "backend:main"

[build-system]
requires = ["uv_build>=0.11.8,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
